feat: move Sequelize config from express to server file
